attendance page: message for future sessions
